chore: rename to .ts
